ELB HTTPS listener config... turning off HTTP
To avoid confusion, I'm turning off HTTP access - see this comment by
Julia:

https://jira.gutools.co.uk/browse/MEM-56?focusedCommentId=68423&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-68423

This follows on from commit b4266f9c5d6. Note that we have to have SSL
certificates registered in AWS with ids that exactly correspond to the
domain-name that they serve, e.g. 'masterclasses-alpha.theguardian.com' or
'membership.theguardian.com'.

The problem of adding the ARN on to the id is neatly addressed by this
stackoverflow answer:

http://stackoverflow.com/a/20012886/438886

...which talks about Cloud Formation pseudo-parameters like
"AWS::AccountId".

http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/pseudo-parameter-reference.html