Remove implicit request from many fragments to remove risk of cache-busting
For pages that aren't signed-in user specific (ie payment, etc) we really
want to make sure the page caches correctly. Passing the request object
through to every single fragment risks that- code might get written that
reads from it!

So, it turns out the request object is consumed in only one place- to
do the return-url for the sign-in link at the top of the page. To satisfy
that, we could pass through some kind of PageContext object (which
basically says where we are) or just default it to
https://membership.theguardian.com/ and enhance with JS...