Use production Guardian Readers events in the CODE env
With the advent of @chrisfinch's Image Uploading Tool, which only
has a Production instance, the CODE site was looking pretty stinky - no
matching images for the other EventBrite account.

We're going to see if we can get by without the effort of doing an
CODE-environment-specific version of the Image Uploading Tool.

EventBrite token