Priority Events presentation code and styling
This commit divides some of @chrisfinch's work and doesn't use the data
from the Event-Ordering tool, instead just taking the first 4 events
returned by Eventbrite as the 'priority' events (the 1st of which is the
'hero' event'). I'll provide another commit to sort that out, but this
commit should allow styling etc to be merged earlier.

https://github.com/guardian/membership/pull/376