Remove AWScala as it's not currently used, use most recent aws-java-sdk
We were using aws-java-sdk v1.7.x, which is pretty out of date. One
consequence of moving to v1.8.x is that the SDK will now pick up your
'default credential profiles file' at ~/.aws/credentials:

http://docs.aws.amazon.com/AWSSdkDocsJava/latest/DeveloperGuide/credentials.html

This seems to me like a nice way to get those AWS credentials into your
JVM, but does mean you gotta be careful that your 'default' entry
is for Membership, and not some other AWS account.


AWScala is pretty cool, but we're not currently using it, so removing it
until we do.

https://github.com/seratch/AWScala