Alway use cached event data... and the member from the request
Apologies, this is two things in one commit, now fiddly for me to untangle,
so here they are, presented together for your delight:

1) As suggested by @wpf500, always use cached event data (rather than
making a call to Eventbrite for the Event Detail pages, etc).

2) When we're calculating the Discount for a Member, don't make a
superfluous call to the Salesforce to get Member details - rely instead on
the Member in the request, already fetched as part of authenticating access
to the page.