Add a friendly page to generate test-usernames [#MEM-447]
This is taking advantage of the Staff-security added by @jennysivapalan,
though we should probably use a different google-group to restrict access
once we've got that sorted out.

Note that I'd like to be logging who requested the generated user-string,
but there seems to be some exisitng weirdness where the authenticated-user
info is not exposed in the resolved type of the request object :-( - need
to get that sorted out before I can log it!

https://jira.gutools.co.uk/browse/MEM-447