Cache 404s on assets for only a short time to mitigate deploy problems
At deploytime, the HTML version served by one appserver may point to CSS
not offered by another appserver. The HTML may be either ahead or behind
of the CSS offered. If it is behind, we should serve-stale (query: is this
actually working). If it is ahead, we don't want to cache the 404 for too
long - as the appservers will soon be in sync again.

A 2-second cache time for a 404 protects us from load very nearly as well
as a 60-second cache time, and problems resolve themselves quicker...

Note that Play's `controllers.Assets` doesn't actually throw a
RuntimeException if the asset falls under a matched path, but with
an unmatched following segment:

https://mem.thegulocal.com/assets/dist/stylesheets/f6d1ede8/style.css

https://mem.thegulocal.com/assets/dist/stylesheets/hamhamham/style.css

Play's `controllers.Assets` would return an (uncached) 404 for the 2nd url,
and so we just need to modify the caching that the response gets. For
successful requests, Play will apply 'sensible' caching already, that we
don't need to modify.