Add diagnostics for the ID-API IdentityServiceErrors we saw today
There was a burst of these at lunchtime today:

```
services.IdentityServiceError: Couldn't find user with ID 1280xxxx
```

The logged error implies that the user doesn't exist in Identity, but that
is not the case. I think we're expecting there to be fields on this user
that do not exist - this logging should make it clearer how the parsing
fails.