Stop Twitter cards derailing users into the mobile apps
Unfortunately the mobile apps don't understand Membership urls, and so
when a user clicks 'Open in the Guardian app', they just get taken to the
'homepage' of the app - not to Membership.

https://cloud.githubusercontent.com/assets/52038/5958317/398506b0-a7bb-11e4-8764-af5d403c8058.gif

While it's nice to promote the apps, until they open the links in a way
that's helpful to the user, we should remove the 'Open in the Guardian app'
functionality from our Twitter cards.

https://dev.twitter.com/cards/mobile

cc @mattandrews