Refactor ticketing model for InternalTicketing & DiscountBenefitTicketing
Some questions only make sense for certain events - eg "What's the member
discount for this ticket?". Events can be broken down like this:

* Internal vs External ticketing (Russell Brannd cinema tickets were
  externally ticketed, by the cinemas involved)
* Events with a discount for certain tiers (Partner, Patron) - which covers
  most events - and events with no discount, like the Charlie Hebdo charity
  event.

There's a lot of conditional logic around how to display this
ticket-pricing information, but rather than relying on devs surrounding
the relevant code with 'if' statements, it's better to use optional
fields and surround with for-comprehensions instead - this way, displaying
`salesDates` for an externally-ticketed event can be enforced as a compile
error - rather than relying on devs to remember to use the right `if`
statement.

As an additional benefit of this work, we actually correctly display
the discount offered based on consideration of both the general-release
ticket and the members-benefit ticket - previously, we were just assuming
that the discount was indeed 20% all the time - I'd rather the discount
displayed correlates with what the user actually sees when they go to
Eventbrite.