Don't show mixed-content warning on past-event pages with related articles
"this page includes other resources which are not secure" - the side box
image which was using old code to pull an image off the related article.

https://membership.theguardian.com/event/dior-and-i-preview-screening-and-director-qa-15946684964

![screenshot](https://cloud.githubusercontent.com/assets/52038/6764621/3cd3adee-cfb1-11e4-8902-02d06b433c50.png)

We've already got [fixed code](https://github.com/guardian/membership-frontend/blob/6f2e9561/frontend/app/model/ResponsiveImage.scala#L15)
for this in `ResponsiveImageGroup`, which converts the image url to the
secure version, so using that fixes the problem and lets us delete some
code too.