Set Cache-Control to serve *stale* cached-content on server errors
For our cached pages (the front of the site, the events pages) this allows
Fastly to serve cached copies of the pages even when the server is down.
This would have been helpful for the outage on Friday 20th February, for
instance.

https://docs.fastly.com/guides/caching/does-fastly-allow-serving-old-content-while-fetching-new-content

With the static parts of the site working even with no operating servers,
we've got to be aware that the uncached parts - generating an eventbrite
discount for ticket purchase, or joining/upgrading as a member, would
still be inoperable and require fixing ASAP.

See also:
https://github.com/guardian/frontend/blob/9205d67/common/app/model/Cached.scala#L39