Load pages of Eventbrite API data in parallel, not sequentially
The old `getPaginated` method was fun (it used `unfoldM`!) but essentially
it was loading one page after another until we got all the data, and was
in any case loading all pages before a result was yielded.

The new method makes an initial call to get the first page of results, then
loads all the other pages concurrently (the 1st call tells it how many
pages of results there will be). As Guardian Live Events & Masterclasses
currently only require 2 pages or less of results each, the performance
behaviour of this code should be identical to before, but it's easier to
understand and should be faster if either account hits 3 or more pages of
data.

The method is now called `getAll` to emphasise the data is not being
lazy-loaded as you traverse over the results.