Inform users that their browser doesn't seem able to handle our Join form
We do client-side validation of form entry, so any error (including the
error of not sending the Stripe token) would indicate that the Javascript
has failed - probably because the browser has Javascript quirks that we
don't see in the modern browsers we test against.

Ideally, we will fix all Javascript issues that unnecessarily stop users
from joining, including adapting ourselves to handle the odd behaviour
of Safari 5.1 on `querySelector()` - but if we /don't/ catch the errors,
we should at least display a helpful message to the user.