Fix staff google-group config so US & Australian staff can join
When we added US and Australian groups in 188b2fb097 we mistakenly gave
the Google Group domains as @theguardian.com - actually, as far as
Google Apps for Business is concerned, the groups are still @guardian.co.uk
so this would never have worked.

This change fixes that issue, but also removes the domain name from the
config altogether, to remove the possibility of similar errors in the
future.