Remove CSP header
We've been running a report-only Content-Security-Policy header on
Membership for the last few weeks. Unfortunately it's pretty hard to
meaningfully determine a safe-whitelist for the policy as the
signal-to-noise ratio on reports is pretty low. We could monitor this
in a smarter way over time (i.e. not using Sentry) but at our current
scale I don't think we get enough benefit out of CSP to be worth the
extra effort. Sadly removing this for now.