Aiming for clarity: create a `findCurrentSubscriptionStatus` method
Following on from feedback that the process to work out the *current*
Zuora subscription version is confusing (which it definitely is),
breakout the logic into it's own heavily commented (boo!) and
descriptively-variable-named (yay!) method.

It's difficult to get this data out of Zuora, this is about as clear as
I can make this process.

https://github.com/guardian/membership-frontend/pull/697#discussion_r38656128