Ensure that we can support "Member's ticket at no extra cost + book"
While the members ticket represents no extra cost to the user, the
Membership Events team often wants to give all users the opportunity to
buy the book associated with the event. The users pick up the book at the
event, but they've paid in advance, thru Eventbrite. Advance purchase
solves a _lot_ of headaches for the Events Team.

This means that there will be some "Member's ticket at no extra cost"
ticket classes which are not zero-cost - because they pay for the book.
They will be called something like:

"Member's ticket at no extra cost + book"

The staff event-overview should show a warning if there is *no* zero-cost
ticket present.