Don't send members who have entered the 'join' flow to downgrade!
If a user is already a member (eg Partner), and they click on a link like
this:

https://membership.theguardian.com/join/supporter/enter-details

(which www.theguardian.com might do, not knowing the status of the member)

...then we *don't* want them to be directed to here:

https://membership.theguardian.com/tier/change/supporter

...because that's a downgrade, that they weren't explicitly expressing
interest in.

If they are hitting join for an _upgrade_ tier (like patron), I think we're
ok to redirect them to:

https://membership.theguardian.com/tier/change/patron

If a user _does_ want to downgrade, they can go directly to

https://membership.theguardian.com/tier/change/friend

...and we don't redirect, but allow the user to proceed as we obviously
should, because they've expressed an explicit interest in downgrading.

https://trello.com/c/yRJfoWrR/198-don-t-take-existing-members-to-tier-downgrade-if-they-clicked-on-a-join-xxxx-enter-details-link