Redirect from 'hidden' events to the replacement events
Following on from the quick evil hack to hide events with bad EventBrite
settings:

https://github.com/guardian/membership-frontend/pull/827

...this addresses a follow-up issue raised by Sophie Davies and
Alice James on the Guardian Live Events team - even though we removed
the events from the Event listing, they're still linked to on Twitter and
from Google- we need to redirect the old pages to the new.

This commit also fixes a problem with the events showing twice on
'filtered' views:

https://membership.theguardian.com/events?location=south-bank

```
---------- Forwarded message ----------
From: Sophie Davies
Date: 26 October 2015 at 17:25
Subject:
To: Alice James

hey alice, the old event pages are showing up in google search - e.g.
this https://membership.theguardian.com/event/guardian-live-the-digital-mirror-old-18882606384

do we know if we can delete them completely?
```