Show default 'Book tickets' message to non-members on Event page
Julia asked for this, on the basis that showing a 'Become a member'
button - but no 'Book tickets' button - to new users (non- members) is
unhelpful. Users are on the event page, and all they really want to do is
book a ticket, so given that they are interested in the event, we should
show them how to proceed from here.

The user will still get taken through the Membership sign-up process before
being able to purchase the ticket.