Soft-deploy session-cam as a switchable feature
Inclusion of the Sessioncam beacon is triggered by the presence of the
'sessioncam-on' cookie. This can be set by staff members by visiting
'/staff/sessioncam/on' and unset by visiting '/staff/sessioncam/off'

We are not adding it by default as it does add an extra 65KB to our JS
build. So we are going to evaluate it first.