Increase uses-per-access-code to 4 for Guardian Live Events
The intention here is to allow users to be able to use 2 'no-extra-cost'
tickets per event instead of 1, which Graham asked if we could do. The code
change is really a bit of a corner-case, the real unlocking of 2
'no-extra-cost' tickets per event is in the Eventbrite admin console,
setting the 'Tickets allowed per order' maximum to `2` instead of `1` for
each event - the Guardian Live Events team will have to pick that up.


The way access codes work is quite subtle. For a given access code:

* The code will unlock specified ticket classes, but does not directly
  allow us to say how many times a ticket in a given class can be used
  with this access code.
* However, the access code does have a "Uses per Offer" (a.k.a
  "quantity_available") attribute, which says how many tickets may be
  bought with it in total.

So, if we want to allow 2 '20% discount' tickets  as well as possibly 2
'no-extra-cost' tickets, the simplest thing to do is just specify a
"quantity_available" value of 4. There is a potential for some
bending-of-the-rules here, which increases as we allow an additional
'no-extra-cost' ticket to be unlocked, but Graham argues that the improved
user experience is worth it.

Note that there is also some accompanying messaging that can be tweaked
in https://github.com/guardian/membership-frontend/blob/383bc16b/frontend/assets/javascripts/src/modules/events/remainingTickets.js#L16-L30
as it currently says, eg, "You can use one of your 6 allocated member
tickets for this event." - this could be refined, but the code will be a
bit more complicated so skipping that for now!