Basic read-only API endpoint to read event image data
Only returns Guardian Live Events, because for the moment that is all
Regis needs - other types may be added in the future.

To get the Grid ImageIdWithCrop into the API, I had to refactor
GridService a bit. The value of `grid.images.media.url` was in the private
config file, but it's a value that's publicly exposed in lots of other
places, so we might as well just have it as a constant - the address of
the grid admin tool is not a secret: "https://media.gutools.co.uk/images/"