Support custom ticket price descriptions on Events
Tina asked if it would be possible to add more explanatory text next to the
ticket price (because the price is quite high due to it including a book)

Eventbrite does allow a 'description' field on tickets:

https://www.eventbrite.co.uk/developer/v3/endpoints/events/#ebapi-post-events-id-ticket-classes

So long as this description field isn't being used for any other purpose,
I think we should be fine to add it as extra descriptive text. Note
that the info must be added to the 'general release' ticket (ie the
first non-hidden ticket).

---------- Forwarded message ----------
From: Tina Moynihan
Date: 2 March 2016 at 17:08
Subject: Adding text?

This event price of £23 and includes a signed book. Can we say this where
I have positioned the yellow arrow? Otherwise the high price point might
put people off from clicking through to the event.

Thanks,
Tina

https://membership.theguardian.com/preview-event/22493299063