Update Salesforce tier on free-to-paid-upgrades
Following on from https://github.com/guardian/members-data-api/pull/95
("Get accurate Membership status from Zuora on webhook update") we noticed
that we were not updating Salesforce with the new Tier when doing a
free-to-paid upgrade (we /do/ do it for all other join/upgrade paths that
I can see). This seemed to stop the Salesforce outgoing webhook from
firing, even though I /think/ Z360 should have taken care of that.

So this change makes the free-to-paid upgrade fall into line
with the other code paths.

Note that it is really sad that we can't just have this bit of code in
*one* place:

```
salesforceService.updateMemberStatus(idUser, tier, customer)
```

...but that's due for a bigger refactor.