Maintenance notice for planned outage due to Salesforce
Salesforce going read-only means we can't register new membership sign-ups
for that period of 2h45m. So, we'll redirect any page that leads to needing
sign-up functionality (eg Supporter Landing Pages, Checkout pages, but not
event pages) to a Maintenance page:

https://membership.theguardian.com/maintenance

This directs people to Contributions:

https://contribute.theguardian.com/?INTCMP=mem_maint

These are the urls that will be affected:

* https://membership.theguardian.com/uk/supporter (all locales)
* https://membership.theguardian.com/join/supporter/enter-details
* All other tiers (Partner, Patron)

As this page needs to go up in the middle of the night (for our
London-based dev team), the `PlannedOutage` class holds the times that
the Maintenance page will be in effect, and for increased confidence that
it's going to work I've added a staff-only page:

https://membership.theguardian.com/planned-outages

Which just says:

```
Next planned outage is
6 November 2016, 2am - 6 November 2016, 4.45am
starting in
38 hours, 1 minute and 40 seconds
```


Salesforce planned outage:


```
RELATED INSTANCES: EU0
RELATED SERVICES: Core Service
PLANNED START TIME: 02:00, Nov 06 GMT
PLANNED END TIME: 04:45, Nov 06 GMT
TYPE: Scheduled Maintenance
AVAILABILITY: This instance will be read only during this maintenance window.
```
https://status.salesforce.com/status/maintenances/9