Refactor config to prepare for deployable .debs & systemd-based AMIs
The config include-chain is now this (each file contains an include for the
next):

- `application.conf`
- `/etc/gu/membership.private.conf` (downloaded from S3)
- `[STAGE].public.conf` (from a stage-specific include in `membership.private.conf`

You can download the private credentials like this:

```
$ aws s3 cp s3://membership-private/DEV/membership.private.conf /etc/gu/membership.private.conf --profile membership
```

This layout of config was adopted for Contributions in July 2016 with
https://github.com/guardian/contributions-frontend/pull/30, and it allows
us to use Debian packages, rather than zips, for our deployable artifacts.
Debian installation doesn't support the `-Dconfig.resource` parameter
we've used with the previous Upstart-based scripts, prompting this
refactor. Also, if we want to use newer Ubuntu AMIs, we need to move to
systemd, which is also supported by the Debian-based artifacts.

Note that `membership.private.conf` is now performing a double-duty:

1. Storing your private credentials!
2. Including the correct public conf for your stage: `[STAGE].public.conf`

The config files for DEV & PROD should be kept very similar in structure,
so that examining them with a diffing tool can easily show any missing keys
or mistakes, ie:

- `DEV.public.conf` & `PROD.public.conf` should diff well
- `DEV/membership.private.conf` & `PROD/membership.private.conf` also