Auto-update AMI with deploy
Following on from https://github.com/guardian/membership-frontend/pull/1376,
now that we're using Amigo AMIs, we can auto-update to the latest AMI with
every deploy.

This mirrors the AMI updating that was added to Contributions with
https://github.com/guardian/contributions-frontend/pull/179 - tho' I've
left out the auto-updating of the cloudformation for the moment.

See the documentation on the `ami-cloudformation-parameter` riffraff
deployment type:

https://riffraff.gutools.co.uk/docs/magenta-lib/types#amicloudformationparameter