Update Nginx setup instructions
Most of the updates come from the fact you need to perform the
`Identity-Platform` setup first, which already gives you Nginx
and a `sites-enabled` folder.

https://github.com/guardian/identity-platform/blob/master/README.md#setup-nginx-for-local-development

The auto-start of nginx with `sudo nginx` removed from `/nginx/setup.sh`
because it actually causes an inconsistent state on Ubuntu 16.10, where the
Nginx PID file can not be created.