Decrease use of `AuthRequest` in the checkout flow
This is a small part of a bigger change working on this story:

https://trello.com/c/TpuX5Fdb/298-a-b-testing-for-new-checkout-flow-test-merging-registration-into-supporter-checkout

To cope with the checkout page no longer being behind an `AuthRequest`,
we need to use `AuthenticatedIdUser` & `AccessCredentials.Cookies`
where in preference to `AuthRequest` where possible in that flow. Note
that `AuthenticatedIdUser` provides a reference to the access crednetials.

Note that this change depends on this update to `identity-play-auth`:

https://github.com/guardian/identity-play-auth/commit/02b40d1ebaf98ecc5ec5f44b8b8480b9409f556a

...that update made it into Membership Frontend with https://github.com/guardian/membership-frontend/pull/1522