Fix race condition while acceptance test is waiting for thank-you
We got a acceptance test failure with this run:

https://saucelabs.com/beta/tests/daab714dab654af19559e3530ae99fd3/commands#210
https://saucelabs.com/beta/tests/daab714dab654af19559e3530ae99fd3/commands#210

The Stripe payment had been accepted, the stripe overlay had just
disappeared, the checkout page was soon to be replaced by the Thank You
page, and the WebDriver code was waiting for the `qa-joiner-summary-tier`
element of the Thank You page to appear. It logged 5 failed attempts to
see the element over a total of 2.5 seconds, each message looking like
this:

```
{"using":"id","value":"qa-joiner-summary-tier"}
RESPONSE

=> {"message":"no such element: Unable to locate element: {\"method\":\"id\",\"selector\":\"qa-joiner-summary-tier\"}\n (Session info: chrome=56.0.2924.76)\n (Driver info: chromedriver=2.27.440174 (e97a722caafc2d3a8b807ee115bfb307f7d2cfd9),platform=Windows NT 6.3.9600 x86_64)"}
```

This is normal, and normally the `WebDriverWait` would persist for up to
30 seconds, waiting for that element to appear. It is configured to swallow
those `org.openqa.selenium.NotFoundException`s:

https://seleniumhq.github.io/selenium/docs/api/java/org/openqa/selenium/support/ui/WebDriverWait.html#WebDriverWait-org.openqa.selenium.WebDriver-long-

However, on the 6th attempt, it got a different error:

```
{"using":"id","value":"qa-joiner-summary-tier"}
RESPONSE

=> {"message":"unknown error: unhandled inspector error: {\"code\":-32000,\"message\":\"Cannot find context with specified id\"}\n (Session info: chrome=56.0.2924.76)\n (Driver info: chromedriver=2.27.440174 (e97a722caafc2d3a8b807ee115bfb307f7d2cfd9),platform=Windows NT 6.3.9600 x86_64)"}
```

This `Cannot find context with specified id` error was propagated,
immediately killing and failing that test.


We believe that the error was caused by the `driver` attempting to look
at the Stripe iframe context, which had just closed (because the test
payment had been accepted) - the context no longer existed, so attempting
to examine items within it was fatal.

We switch the context of the Stripe iframe earlier in test, in order to
be able to fill the fields out:

https://github.com/guardian/membership-frontend/blob/940fe2a95/frontend/test/acceptance/JoinSupporterSpec.scala#L81-L103

...up until now, we had not explicitly switched back. We probably didn't
need to, because the page-load of the Thank You page was probably enough
to switch driver context automatically.

This change explicitly switches the context of the driver to the default
content, so that we don't attempt to operate on the Stripe iframe which
is about to disappear.

Also fixing small typo of `waitTimout`.