Fix PayPal payments on the not-yet-Identity-authenticated flow
This follows on from #1589, where Santiago discovered we couldn't get
PayPal payments to complete:

https://github.com/guardian/membership-frontend/pull/1589#issuecomment-293890744

This change removes the requirement that the user is authenticated to hit
these endpoints:

* `POST /paypal/setup-payment`
* `POST /paypal/create-agreement`

This means they can hit the PayPal button before we've created them an
Identity account. The main bit of functionality this affects is Test-Users,
as the logged-in user was used to determine which PayPal backend to use.
With this change, the `pre-signin-test-user` cookie (dropped by visiting
https://membership.theguardian.com/test-users) is now permitted as an
alternative way of selecting the Test backend.

There's also some small tidy-ups here, using explicit body parsers to
avoid having to do the Json parsing/validation ourselves:

https://www.playframework.com/documentation/2.5.x/ScalaBodyParsers#Choosing-an-explicit-body-parser


https://trello.com/c/TpuX5Fdb/298-a-b-testing-for-new-checkout-flow-test-merging-registration-into-supporter-checkout