Add receive-gnm-marketing checkbox
Spoke to Evie Wilson & Robert Chilvers on the Data Privacy team regarding
the 'Receive Guardian Marketing' checkbox. They are happy for us to go
ahead with the current wording as used on
https://profile.theguardian.com/register, ie:

    "Keep me up to date with offers from the Guardian"

Evie said it would be best to put the checkbox near where we're asking for
the email, to make it clear what it was in relation to (so near the start
of the flow, rather than the end, in our case).

The checkbox must default to 'off' for new users, this is a requirement of
GDPR.

They pointed out that General Data Protection Regulation (GDPR) is coming
into effect in 2018, at which point we will be required to return an audit
trail, with each change being accompanied by:

* DateTime
* Application where change was made (eg Identity or Membership)
* What wording was used when the user accepted the mailing

...however, that is going to be a larger project which we do not need to
implement now.

https://trello.com/c/TpuX5Fdb/298-a-b-testing-for-new-checkout-flow-test-merging-registration-into-supporter-checkout