Hard-code `frontend` here - it's not actually set by the AWS ${App} var
The reason that first `frontend` in `/var/log/frontend/frontend.log` is
actually `frontend` and not something else has nothing to do with the AWS
`${App}` cloud-formation variable - it's set by the Debian package, and
that in turn is a consequence of the sbt configuration. So using the ${App}
is probably not helpful.

https://github.com/guardian/membership-frontend/pull/1610#discussion_r114519478