Fix AWS CloudFormation template validation error on `Fn::Sub`
```
Template error: One or more Fn::Sub intrinsic functions don't specify
expected arguments. Specify a string as first argument, and an optional
second argument to specify a mapping of values to replace in the string
```

I needed to add `App` as a template parameter in order to use it
both in the cloud-init script and tagging.