Remove vestiges of password-strength meter
The password strength meter (powered by https://github.com/dropbox/zxcvbn)
was broken, and causing JS errors on the console:

```
Failed to load resource: the server responded with a status of 404 (Not Found)
curl.js:6 Uncaught Error: Syntax or http error: /assets/javascripts/lib/zxcvbn/zxcvbn.js
    at HTMLScriptElement.c.onerror (curl.js:14)
```

https://sentry.io/the-guardian/membership-client-side/issues/90901522/

I had a chat with Mark Butler on the Identity team about this, and the fact
that https://profile.theguardian.com/register no longer shows a password
strength meter, and he told me that password-strength-meters are now
regarded as a bit of a bad idea, providing a false sense of security.

See also:

https://nakedsecurity.sophos.com/2016/08/17/why-you-still-cant-trust-password-strength-meters/
https://security.stackexchange.com/questions/51060/should-i-reject-obviously-poor-passwords

...of the available password-strength-meters, `zxcvbn` seems to be the best
one, but it seems it's still too easy for weak passwords to slip by.