Fix "You're signed in as ..." message on checkout for mobile-app users
This changes the approach used to populate username in the "You're signed
in as Jess Blogs" message from client-side to server-side, as the
checkout page (https://membership.theguardian.com/join/supporter/enter-details)
already does an Identity API request to get all user details before
rendering.

The old client-side approach depended on the `GU_U` cookie to get the
display-name, which has so far not been available to mobile-apps. The apps
_mainly_ just use an access token, not cookies, but have been able to
exchange that access token for an `SC_GU_U` cookie (*not* the `GU_U`
cookie which contains the display-name) since https://github.com/guardian/identity/pull/553

In the past day or two, there was work done with
https://github.com/guardian/identity/pull/663 to also give a `GU_U` cookie
in return for the token, but that was reverted with
https://github.com/guardian/identity/pull/664 when we realised doing the
fix server-side in membership-frontend was a better approach.